Move API utils into dedicated apis directory
