+++ apps/website/src/server/apis/weather.ts
